requirements.txt
###### Libraries to install ######
cairosvg
Pillow